+++ packages/providers/lib/json-rpc-provider.js
@@ -232,18 +232,45 @@ var JsonRpcSigner = /** @class */ (function (_super) {
         });
     };
     JsonRpcSigner.prototype.sendTransaction = function (transaction) {
-        var _this = this;
-        return this.sendUncheckedTransaction(transaction).then(function (hash) {
-            return web_1.poll(function () {
-                return _this.provider.getTransaction(hash).then(function (tx) {
-                    if (tx === null) {
-                        return undefined;
-                    }
-                    return _this.provider._wrapTransaction(tx, hash);
-                });
-            }, { oncePoll: _this.provider }).catch(function (error) {
-                error.transactionHash = hash;
-                throw error;
+        return __awaiter(this, void 0, void 0, function () {
+            var blockNumber, hash, error_1;
+            var _this = this;
+            return __generator(this, function (_a) {
+                switch (_a.label) {
+                    case 0: return [4 /*yield*/, this.provider._getInternalBlockNumber(100 + 2 * this.provider.pollingInterval)];
+                    case 1:
+                        blockNumber = _a.sent();
+                        return [4 /*yield*/, this.sendUncheckedTransaction(transaction)];
+                    case 2:
+                        hash = _a.sent();
+                        _a.label = 3;
+                    case 3:
+                        _a.trys.push([3, 5, , 6]);
+                        return [4 /*yield*/, web_1.poll(function () { return __awaiter(_this, void 0, void 0, function () {
+                                var tx;
+                                return __generator(this, function (_a) {
+                                    switch (_a.label) {
+                                        case 0: return [4 /*yield*/, this.provider.getTransaction(hash)];
+                                        case 1:
+                                            tx = _a.sent();
+                                            if (tx === null) {
+                                                return [2 /*return*/, undefined];
+                                            }
+                                            return [2 /*return*/, this.provider._wrapTransaction(tx, hash, blockNumber)];
+                                    }
+                                });
+                            }); }, { oncePoll: this.provider })];
+                    case 4: 
+                    // Unfortunately, JSON-RPC only provides and opaque transaction hash
+                    // for a response, and we need the actual transaction, so we poll
+                    // for it; it should show up very quickly
+                    return [2 /*return*/, _a.sent()];
+                    case 5:
+                        error_1 = _a.sent();
+                        error_1.transactionHash = hash;
+                        throw error_1;
+                    case 6: return [2 /*return*/];
+                }
             });
         });
     };
@@ -395,7 +422,7 @@ var JsonRpcProvider = /** @class */ (function (_super) {
     };
     JsonRpcProvider.prototype._uncachedDetectNetwork = function () {
         return __awaiter(this, void 0, void 0, function () {
-            var chainId, error_1, error_2, getNetwork;
+            var chainId, error_2, error_3, getNetwork;
             return __generator(this, function (_a) {
                 switch (_a.label) {
                     case 0: return [4 /*yield*/, timer(0)];
@@ -410,7 +437,7 @@ var JsonRpcProvider = /** @class */ (function (_super) {
                         chainId = _a.sent();
                         return [3 /*break*/, 9];
                     case 4:
-                        error_1 = _a.sent();
+                        error_2 = _a.sent();
                         _a.label = 5;
                     case 5:
                         _a.trys.push([5, 7, , 8]);
@@ -419,7 +446,7 @@ var JsonRpcProvider = /** @class */ (function (_super) {
                         chainId = _a.sent();
                         return [3 /*break*/, 8];
                     case 7:
-                        error_2 = _a.sent();
+                        error_3 = _a.sent();
                         return [3 /*break*/, 8];
                     case 8: return [3 /*break*/, 9];
                     case 9:
@@ -548,7 +575,7 @@ var JsonRpcProvider = /** @class */ (function (_super) {
     };
     JsonRpcProvider.prototype.perform = function (method, params) {
         return __awaiter(this, void 0, void 0, function () {
-            var args, error_3;
+            var args, error_4;
             return __generator(this, function (_a) {
                 switch (_a.label) {
                     case 0:
@@ -562,8 +589,8 @@ var JsonRpcProvider = /** @class */ (function (_super) {
                         return [4 /*yield*/, this.send(args[0], args[1])];
                     case 2: return [2 /*return*/, _a.sent()];
                     case 3:
-                        error_3 = _a.sent();
-                        return [2 /*return*/, checkError(method, error_3, params)];
+                        error_4 = _a.sent();
+                        return [2 /*return*/, checkError(method, error_4, params)];
                     case 4: return [2 /*return*/];
                 }
             });